fix up html
