+++ packages/web/src/App.css
@@ -19,34 +19,3 @@ h1 {
   margin-bottom: 5vh; /* 5% of the viewport height as margin bottom */
   font-size: 5vw; /* Font size is 5% of the viewport width */
 }
-
-.calendar-container {
-  width: 80vw; /* 80% of the viewport width */
-  max-width: 800px;
-  border: 1px solid #ccc; /* Lighter border color */
-  padding: 5vw; /* 5% of the viewport width as padding */
-  border-radius: 3vw; /* 3% of the viewport width as border radius */
-  background-color: #fff; /* White background color for the calendar container */
-  box-shadow: 0 0 2vw rgba(0, 0, 0, 0.1); /* Light shadow for the container */
-}
-
-.rbc-calendar {
-  height: 40vh; /* 40% of the viewport height */
-  background-color: #f5f5f5; /* Lighter background color for the calendar */
-}
-
-.rbc-timeslot-group {
-  border-bottom: 0.5vh solid #ddd; /* Lighter border color */
-}
-
-.rbc-event {
-  background-color: #4caf50; /* Event background color */
-  color: #fff;
-  padding: 1vw 2vw; /* Padding is 1% of the viewport width and 2% of the viewport width */
-  border-radius: 1vw; /* Border radius is 1% of the viewport width */
-  cursor: pointer;
-}
-
-.rbc-selected-cell {
-  background-color: rgba(0, 123, 255, 0.1); /* Selected time slot background color */
-}
