+++ packages/web/src/components/Header/Header.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export const Header = () => (
+    <div>
+        <h1>Pickleball Scheduler</h1>
+        <div>For the boys at Van Dyke Park</div>
+    </div>
+);
